refactor: consolidate the metrics domain and move the websocket config #307
