+++ backend/src/main/java/com/occupi/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.occupi.feature.receiver;
+package com.occupi.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
