+++ backend/src/main/java/com/occupi/feature/metrics/MetricsService.java
@@ -1,7 +1,5 @@
-package com.occupi.feature.database.service;
+package com.occupi.feature.metrics;
 
-import com.occupi.feature.database.model.MetricsData;
-import com.occupi.feature.database.repository.MetricsRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
