+++ backend/src/main/java/com/occupi/feature/metrics/PiMetricsService.java
@@ -1,0 +1,7 @@
+package com.occupi.feature.metrics;
+
+import com.occupi.feature.metrics.dto.Metrics;
+
+public interface PiMetricsService {
+    void process(Metrics metrics);
+}
