--- backend/src/main/java/com/occupi/feature/receiver/PiMetricsService.java
@@ -1,7 +1,0 @@
-package com.occupi.feature.receiver;
-
-import com.occupi.feature.receiver.dto.Metrics;
-
-public interface PiMetricsService {
-    void process(Metrics metrics);
-}
