+++ backend/src/test/java/com/occupi/feature/occupancy/LastCacheInfluxIntegrationTest.java
@@ -4,8 +4,8 @@ import com.influxdb.v3.client.InfluxDBClient;
 import com.influxdb.v3.client.query.QueryOptions;
 import com.occupi.config.InfluxDBProperties;
 import com.occupi.config.InfluxLastCacheInitializer;
-import com.occupi.feature.database.model.MetricsData;
-import com.occupi.feature.database.repository.MetricsRepository;
+import com.occupi.feature.metrics.MetricsData;
+import com.occupi.feature.metrics.MetricsRepository;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
